+++ src/components/add-monster/MonsterSearch.js
@@ -1,6 +1,5 @@
 import { useState, useRef, useEffect } from "react"
 import Character from 'models/Character.js'
-import { useCharactersContext } from "context/characters-context"
 import { getMonster } from "services/dnd-5e-api"
 import { getAllMonsters } from "services/dnd-5e-api"
 
@@ -56,7 +55,7 @@ export default function MonsterSearch({additionalMonsters, setMessage, addCharac
         </datalist>
 
         <input type="submit"
-        value={"Search"}/>
+        value="Add Monster"/>
       </div>
 
     </form>